Spin-off penbox as its own service
