+++ meerk40t/core/node/op_hatch.py
@@ -316,7 +316,7 @@ class HatchOpNode(Node, Parameters):
             penbox_pass = self.settings.get("penbox_pass")
             if penbox_pass is not None:
                 try:
-                    penbox_pass = context.elements.penbox[penbox_pass]
+                    penbox_pass = context.penbox.penbox[penbox_pass]
                 except KeyError:
                     penbox_pass = None
             hatch_cache = dict()
